requirements.txt
selenium == 4.20.0
robotframework == 7.0
robotframework-seleniumlibrary == 6.3.0